Generalize the chaincode language type supported
Things are somewhat hardcoded to only support GOLANG.  Lets generalize this.
