+++ openchain/devops.go
@@ -263,13 +263,15 @@ func CheckSpec(spec *pb.ChaincodeSpec) error {
 		return errors.New("Expected chaincode specification, nil received")
 	}
 
-	// Only allow GOLANG type at the moment
-	if spec.Type != pb.ChaincodeSpec_GOLANG {
-		return fmt.Errorf("Only support '%s' currently", pb.ChaincodeSpec_GOLANG)
+	switch spec.Type {
+	case pb.ChaincodeSpec_GOLANG:
+		if err := checkGolangSpec(spec); err != nil {
+			return err
+		}
+	default:
+		return fmt.Errorf("Unsupported spec: %s", spec.Type)
 	}
-	if err := checkGolangSpec(spec); err != nil {
-		return err
-	}
+
 	devopsLogger.Debug("Validated spec:  %v", spec)
 
 	return nil
